[dawn][native] Spanify RenderPassEncoder::ExecuteBundles
Adds more spanification TODOs.

Changes ExecuteBundleCmd's count to be a size_t to match the
renderBundleCount argument.

Bug: 515272358, 512465980, 512465980, 528305452
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/320176
Commit-Queue: Corentin Wallez <cwallez@chromium.org>

+++ src/dawn/native/Commands.h
@@ -332,7 +332,7 @@ struct EndRenderPassCmd {
 };
 
 struct ExecuteBundlesCmd {
-    uint32_t count = 0;
+    size_t count = 0;
 };
 
 struct ClearBufferCmd {
